[skip actions] lab/lab01/ch01_t10_two_types_of_division.py

+++ lab/lab01/ch01_t10_two_types_of_division.py
@@ -5,4 +5,4 @@ whole_cucumbers_per_perseom - cucumbers // num_people
 print(whole_cucumbers_per_person)
 
 float_cucumbers / num_people
-print(float)
+print(float_)
